fix: fix parallel builds of the same app from the same machine

+++ src/services/common/git-service.js
@@ -2,20 +2,24 @@ const path = require("path");
 const _ = require("lodash");
 
 class GitService {
-    constructor($childProcess, $fs, $logger, $cleanupService, gitDirsPath, gitRepoName, workingDirectory) {
+    constructor($childProcess, $fs, $logger, $cleanupService, gitDirsPath, gitRepoName, workingDirectory, uniqueId) {
         this.$childProcess = $childProcess;
         this.$fs = $fs;
         this.$logger = $logger;
         this.$cleanupService = $cleanupService;
         this.gitDirsPath = gitDirsPath;
-        this.gitRepoName = gitRepoName;
+        this.uniqueId = uniqueId;
+        this.gitRepoName = gitRepoName + "-" + uniqueId;
+        this.defaultBranchName = GitService.TEMP_BRANCH_NAME_PREFIX + uniqueId;
         this.workingDirectory = workingDirectory;
     }
 
-    async gitPushChanges(remoteUrl, mappedFiles, placeholders, branchId) {
-        // a workaround for a sporadic "git exited with code 1"
-        this.cleanGitRepository();
+    async gitPushChanges(remoteUrl, mappedFiles, placeholders) {
+        // ensure clean repo as a workaround for a sporadic "git exited with code 1"
+        await this.cleanGitRepository();
         await this.gitInit();
+        // clean on ctrl + c
+        await this.cleanGitRepository(true);
         const isRemoteAdded = await this.gitCheckIfRemoteIsAdded(GitService.REMOTE_NAME);
         if (isRemoteAdded) {
             const isGitRemoteCorrect = await this.isGitRemoteSetToCorrectUrl(remoteUrl);
@@ -34,13 +38,13 @@ class GitService {
             await this.executeCommand(["config", "--local", "core.autocrlf", "false"]);
         }
 
-        await this.checkoutBranch(GitService.TEMP_BRANCH_NAME_PREFIX + branchId);
+        await this.checkoutBranch();
         const statusResult = await this.gitStatus();
         this.$logger.trace(`Result of git status: ${statusResult}.`);
         let revision;
         if (this.hasNothingToCommit(statusResult.stdout)) {
             this.$logger.trace("Nothing to commit. Just push force the branch.");
-            await this.gitPush(branchId);
+            await this.gitPush();
             revision = await this.getCurrentRevision();
             return revision;
         }
@@ -71,39 +75,41 @@ class GitService {
         }
 
         await this.gitCommit();
-        await this.gitPush(branchId);
+        await this.gitPush();
         revision = await this.getCurrentRevision();
         return revision;
     }
 
-    async checkoutBranch(branchId, cleanUpOnly) {
+    async checkoutBranch(branchName, cleanUpOnly) {
+        branchName = branchName || this.defaultBranchName;
         try {
-            await this.executeCommand(["checkout", branchId], undefined, false, false, cleanUpOnly);
+            await this.executeCommand(["checkout", branchName], undefined, false, false, cleanUpOnly);
         }
         catch (error) {
-            await this.executeCommand(["checkout", "-b", branchId], undefined, false, false, cleanUpOnly);
-            const tempBranchPrefixIndex = branchId.indexOf(GitService.TEMP_BRANCH_NAME_PREFIX);
+            await this.executeCommand(["checkout", "-b", branchName], undefined, false, false, cleanUpOnly);
+            const tempBranchPrefixIndex = branchName.indexOf(GitService.TEMP_BRANCH_NAME_PREFIX);
             if (!cleanUpOnly && tempBranchPrefixIndex === 0) {
-                const buildId = branchId.substring(GitService.TEMP_BRANCH_NAME_PREFIX.length);
-                await this.gitDeleteTempBuildBranch(buildId, true);
+                await this.gitDeleteBranch(branchName, true);
             }
         }
     }
 
-    async gitDeleteTempBuildBranch(cliBuildId, cleanUpOnly) {
+    async gitDeleteBranch(branchName, cleanUpOnly) {
+        branchName = branchName || this.defaultBranchName;
         // in order to delete the local branch, we need to change the checked out one
         await this.checkoutBranch("master", cleanUpOnly);
         // remove local branch
-        await this.executeCommand(["branch", "-D", GitService.TEMP_BRANCH_NAME_PREFIX + cliBuildId], undefined, false, false, cleanUpOnly)
+        await this.executeCommand(["branch", "-D", branchName], undefined, false, false, cleanUpOnly)
         // remove remote branch
         const env = _.assign({}, process.env);
-        return this.executeCommand(["push", "-d", GitService.REMOTE_NAME, GitService.TEMP_BRANCH_NAME_PREFIX + cliBuildId], { env, cwd: this.workingDirectory }, false, false, cleanUpOnly);
+        return this.executeCommand(["push", "-d", GitService.REMOTE_NAME, branchName], { env, cwd: this.workingDirectory }, false, false, cleanUpOnly);
     }
 
     async getCurrentRevision() {
         const revisionCommandResult = await this.executeCommand(["rev-parse", "HEAD"]);
         return revisionCommandResult.stdout.trim();
     }
+
     async isGitRemoteSetToCorrectUrl(remoteUrl) {
         const result = await this.executeCommand(["remote", "-v"]);
         return result.stdout.indexOf(remoteUrl.httpRemoteUrl) !== -1;
@@ -132,10 +138,12 @@ class GitService {
     async gitStatus() {
         return this.executeCommand(["status"]);
     }
-    async gitPush(cliBuildId) {
+    async gitPush(branchName) {
+        branchName = branchName || this.defaultBranchName;
         const env = _.assign({}, process.env);
-        return this.executeCommand(["push", "--force", GitService.REMOTE_NAME, GitService.TEMP_BRANCH_NAME_PREFIX + cliBuildId], { env, cwd: this.workingDirectory });
+        return this.executeCommand(["push", "--force", GitService.REMOTE_NAME, branchName], { env, cwd: this.workingDirectory });
     }
+
     async gitRemoteAdd(remoteUrl) {
         return this.executeCommand(["remote", "add", GitService.REMOTE_NAME, remoteUrl.httpRemoteUrl]);
     }
@@ -173,13 +181,14 @@ class GitService {
             return this.$childProcess.spawnFromEvent("git", args, "close", options, spawnFromEventOptions);
         }
     }
-    isGitRepository() {
-        return this.$fs.exists(this.getGitDirPath());
-    }
-    cleanGitRepository() {
+    async cleanGitRepository(cleanUpOnly) {
         const gitDir = this.getGitDirPath();
-        if (this.$fs.exists(gitDir)) {
-            this.$fs.deleteDirectory(gitDir);
+        if (cleanUpOnly) {
+            await this.$cleanupService.addCleanupDeleteAction(gitDir);
+        } else {
+            if (this.$fs.exists(gitDir)) {
+                this.$fs.deleteDirectory(gitDir);
+            }
         }
     }
     getGitDirPath() {
